requirements/base.txt


# Bleeding edge Django
django==1.9.9

# Configuration
django-environ==0.4.0

# Database driver
mysqlclient==1.3.7

# Forms
django-braces==1.9.0
django-crispy-forms==1.6.0

# Models
django-model-utils==2.5.2

# Images
Pillow==3.3.1

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.27.0

# Unicode slugification
awesome-slugify==1.6.5
django-autoslug==1.9.3

# Time zones support
pytz==2016.6.1

# Redis support
django-redis==4.4.4
redis>=2.10.0
cached-property==1.3.0
django-import-export==0.5.0
django-bootstrap-pagination==1.6.2
